--- packages/ordered-list/src/ListItem/index.ts
@@ -1,2 +1,0 @@
-export { ListItem } from './ListItem';
-export { ListItemProps } from './ListItem.types';
