+++ packages/ordered-list/src/OrderedListItem/index.ts
@@ -1,0 +1,2 @@
+export { OrderedListItem } from './OrderedListItem';
+export { OrderedListItemProps } from './OrderedListItem.types';
